% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifmatchr.R
\docType{data}
\name{example_motifs}
\alias{example_motifs}
\title{example_motifs}
\usage{
data(example_motifs)
}
\value{
\code{\link[TFBSTools]{PFMatrixList}} of length 3
}
\description{
A few example motifs from JASPAR 2016 for trying out motifmatchr
}
\examples{

data(example_motifs, package = "motifmatchr")

}
\keyword{datasets}
