% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\docType{methods}
\name{motif_counts}
\alias{motif_counts}
\alias{motif_counts,SummarizedExperiment-method}
\title{motif_counts}
\usage{
motif_counts(object)

\S4method{motif_counts}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment object with counts assay}
}
\value{
matrix with counts
}
\description{
get motif counts from SummarizedExperiment object
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment
}}

\examples{

data(example_motifs, package = "motifmatchr")

# Make a set of peaks
peaks <- GRanges(seqnames = c("chr1","chr2","chr2"),
                ranges = IRanges(start = c(76585873,42772928,100183786),
                                 width = 500))

# Get motif matches for example motifs  (using hg19 genome, the default)
motif_ix <- match_motifs(example_motifs, peaks, out = "scores")

motif_counts(motif_ix)
}
