% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pwms.R
\docType{methods}
\name{motif_matches,SummarizedExperiment-method}
\alias{motif_matches,SummarizedExperiment-method}
\title{motif_matches}
\usage{
\S4method{motif_matches}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment object with matches assay}
}
\value{
matrix with scores
}
\description{
get motif matches from SummarizedExperiment object
}

