% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pwms.R
\docType{methods}
\name{motif_scores,SummarizedExperiment-method}
\alias{motif_scores,SummarizedExperiment-method}
\title{motif_scores}
\usage{
\S4method{motif_scores}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment object with scores assay}
}
\value{
matrix with scores
}
\description{
get motif scores from SummarizedExperiment object
}

