% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\name{motif_scores}
\alias{motif_scores}
\title{motif_scores}
\usage{
motif_scores(object)
}
\arguments{
\item{object}{SummarizedExperiment object with scores assay}
}
\value{
matrix with scores
}
\description{
get motif scores from SummarizedExperiment object
}

