% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifmatchr.R
\docType{package}
\name{motifmatchr}
\alias{motifmatchr}
\alias{motifmatchr-package}
\title{motifmatchr: Fast Motif Matching in R}
\description{
The motifmatchr package is designed for analyzing many sequences and many
motifs to find which sequences contain which motifs.
}
\details{
motifmatchr uses the MOODS C++ library (developedby Pasi Rastas,
Janne Korhonen, and Petri Martinmaki) internally for motif matching.

The primary method of motifmatchr is \code{\link{matchMotifs}}, which takes
in motif PWMs/PFMs and genomic ranges or sequences and returns either which
ranges/sequences match which motifs or the positions of the matches.

Compared with alternative motif matching functions available in Bioconductor
(e.g. matchPWM in Biostrings or searchSeq in TFBSTools), motifmatchr is
designed specifically for the use case of determining whether many different
sequences/ranges contain many different motifs.
}
\author{
Alicia Schep
}
