% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{motifmatchr_deprecated}
\alias{motifmatchr_deprecated}
\alias{motif_matches}
\alias{motif_counts}
\alias{motif_scores}
\alias{match_motifs}
\title{Deprecated functions in motifmatchr}
\usage{
motif_matches(...)

motif_counts(...)

motif_scores(...)

match_motifs(...)
}
\arguments{
\item{...}{arguments passed to new function}
}
\value{
calls the replacement function
}
\description{
motifmatchr has moved functions and methods to camelCase from snake_case.
The following functions have been deprecated and replaced with a different
name:
\itemize{
  \item motif_matches is now \code{\link{motifMatches}}
  \item motif_counts is now \code{\link{motifCounts}}
  \item motif_scores is now  \code{\link{motifScores}}
  \item match_motifs is now  \code{\link{matchMotifs}}
}
}
\author{
Alicia Schep
}
