% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_pwms_utils.R, R/utils.R
\name{pwm_type}
\alias{pwm_type}
\title{pwm_type}
\usage{
pwm_type(pwm)

pwm_type(pwm)
}
\arguments{
\item{pwm}{PWMatrix object}

\item{PWMatrix}{object}
}
\value{
'log','log2', or 'frequency' depending on type of pwm

'log','log2', or 'frequency' depending on type of pwm
}
\description{
pwm_type

Determines type of PWM
}

