% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifmatchr.R
\docType{data}
\name{example_motifs}
\alias{example_motifs}
\title{example_motifs}
\format{\code{\link[TFBSTools]{PFMatrixList}} of length 3}
\usage{
data(example_motifs)
}
\description{
A few example motifs from JASPAR 2016 for trying out motifmatchr
}
\keyword{datasets}

