% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\docType{methods}
\name{match_motifs}
\alias{match_motifs}
\alias{match_motifs,PWMatrixList,DNAStringSet-method}
\alias{match_motifs,PWMatrixList,character-method}
\alias{match_motifs,PWMatrixList,DNAString-method}
\alias{match_motifs,PWMatrixList,GenomicRanges-method}
\alias{match_motifs,PWMatrixList,RangedSummarizedExperiment-method}
\alias{match_motifs,PFMatrixList,ANY-method}
\alias{match_motifs,PWMatrix,ANY-method}
\alias{match_motifs,PFMatrix,ANY-method}
\title{match_motifs}
\usage{
match_motifs(pwms, subject, ...)

\S4method{match_motifs}{PWMatrixList,DNAStringSet}(pwms, subject,
  genome = NULL, bg = NULL, out = c("matches", "scores", "positions"),
  p.cutoff = 5e-05, w = 7, ranges = NULL)

\S4method{match_motifs}{PWMatrixList,character}(pwms, subject, genome = NULL,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_motifs}{PWMatrixList,DNAString}(pwms, subject, genome = NULL,
  bg = NULL, out = c("matches", "scores", "positions"), p.cutoff = 5e-05,
  w = 7, ranges = NULL)

\S4method{match_motifs}{PWMatrixList,GenomicRanges}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19, bg = NULL, out = c("matches",
  "scores", "positions"), p.cutoff = 5e-05, w = 7, ranges = NULL)

\S4method{match_motifs}{PWMatrixList,RangedSummarizedExperiment}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19, bg = NULL, out = c("matches",
  "scores", "positions"), p.cutoff = 5e-05, w = 7, ranges = NULL)

\S4method{match_motifs}{PFMatrixList,ANY}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19, bg = NULL, out = c("matches",
  "scores", "positions"), p.cutoff = 5e-05, w = 7, ranges = NULL)

\S4method{match_motifs}{PWMatrix,ANY}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19, bg = NULL, out = c("matches",
  "scores", "positions"), p.cutoff = 5e-05, w = 7, ranges = NULL)

\S4method{match_motifs}{PFMatrix,ANY}(pwms, subject,
  genome = BSgenome.Hsapiens.UCSC.hg19, bg = NULL, out = c("matches",
  "scores", "positions"), p.cutoff = 5e-05, w = 7, ranges = NULL)
}
\arguments{
\item{pwms}{either \code{\link[TFBSTools]{PFMatrix}},
\code{\link[TFBSTools]{PFMatrixList}}, \code{\link[TFBSTools]{PWMatrix}},
\code{\link[TFBSTools]{PWMatrixList}}}

\item{subject}{either \code{\link[GenomicRanges]{GenomicRanges}},
\code{\link[Biostrings]{DNAStringSet}}, \code{\link[Biostrings]{DNAString}},
or character vector}

\item{...}{additional arguments depending on inputs}

\item{genome}{BSgenome object, only used if subect is
\code{\link[GenomicRanges]{GenomicRanges}}}

\item{bg}{background nucleotide frequencies. if not provided, computed from
subject}

\item{out}{what to return? see return section}

\item{p.cutoff}{p-value cutoff for returning motifs}

\item{w}{parameter controlling size of window for filtration; default is 7}

\item{ranges}{if subject is not GenomicRanges, ranges to use when out is
positions}
}
\value{
Either returns a SummarizedExperiment with a sparse matrix with
 values set to TRUE for a match (if return == 'matches'), a
 SummarizedExperiment with a matches matrix as well as matrices with the
 maximum motif score and total motif counts (if return == 'scores'), or a
 GenomicRanges or IRanges object with all the positions of matches or
 \code{\link[GenomicRanges]{GenomicRanges}} if positions
}
\description{
Find motif matches
}
\section{Methods (by class)}{
\itemize{
\item \code{pwms = PWMatrixList,subject = DNAStringSet}: PWMatrixList/DNAStringSet

\item \code{pwms = PWMatrixList,subject = character}: PWMatrixList/character

\item \code{pwms = PWMatrixList,subject = DNAString}: PWMatrixList/DNAString

\item \code{pwms = PWMatrixList,subject = GenomicRanges}: PWMatrixList/GenomicRanges

\item \code{pwms = PWMatrixList,subject = RangedSummarizedExperiment}: PWMatrixList/RangedSummarizedExperiment

\item \code{pwms = PFMatrixList,subject = ANY}: PFMatrixList/ANY

\item \code{pwms = PWMatrix,subject = ANY}: PWMatrix/ANY

\item \code{pwms = PFMatrix,subject = ANY}: PFMatrix/ANY
}}

\examples{

data(example_motifs, package = "motifmatchr")

# Make a set of peaks
peaks <- GRanges(seqnames = c("chr1","chr2","chr2"),
                ranges = IRanges(start = c(76585873,42772928,100183786),
                                 width = 500))

# Get motif matches for example motifs  (using hg19 genome, the default)
motif_ix <- match_motifs(example_motifs, peaks)

}
