% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\name{motif_counts}
\alias{motif_counts}
\title{motif_counts}
\usage{
motif_counts(object)
}
\arguments{
\item{object}{SummarizedExperiment object with counts assay}
}
\value{
matrix with counts
}
\description{
get motif counts from SummarizedExperiment object
}

