% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\docType{methods}
\name{motif_counts}
\alias{motif_counts}
\alias{motif_counts,SummarizedExperiment-method}
\title{motif_counts}
\usage{
motif_counts(object)

\S4method{motif_counts}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment object with counts assay}
}
\value{
matrix with counts
}
\description{
get motif counts from SummarizedExperiment object
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment
}}

