% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\docType{methods}
\name{motif_scores}
\alias{motif_scores}
\alias{motif_scores,SummarizedExperiment-method}
\title{motif_scores}
\usage{
motif_scores(object)

\S4method{motif_scores}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment object with scores assay}
}
\value{
matrix with scores
}
\description{
get motif scores from SummarizedExperiment object
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment
}}

