% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_motifs.R
\docType{methods}
\name{motif_scores}
\alias{motif_scores}
\alias{motif_scores,SummarizedExperiment-method}
\title{motif_scores}
\usage{
motif_scores(object)

\S4method{motif_scores}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{SummarizedExperiment object with scores assay}
}
\value{
matrix with scores
}
\description{
get motif scores from SummarizedExperiment object
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: method for SummarizedExperiment
}}

\examples{

data(example_motifs, package = "motifmatchr")

# Make a set of peaks
peaks <- GenomicRanges::GRanges(seqnames = c("chr1","chr2","chr2"),
                ranges = IRanges::IRanges(start = c(76585873,42772928,
                                          100183786),
                                          width = 500))

# Get motif matches for example motifs
motif_ix <- match_motifs(example_motifs, peaks, genome = "hg19",
                         out = "scores")

motif_scores(motif_ix)
}
